% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package_documentation.R
\docType{package}
\name{asymLD-package}
\alias{asymLD}
\alias{asymLD-package}
\title{asymLD: a package for computing asymmetric Linkage Disequilibrium (ALD) between two polymorphic genetic loci}
\description{
Computes asymmetric LD measures (ALD) for polymorphic genetic data.
These measures are identical to the correlation measure (r) for bi-allelic data.
}
\details{
\tabular{ll}{
 Package: \tab asymLD\cr
 Type: \tab Package\cr
 Version: \tab 0.1\cr
 Date: \tab 2015-03-13\cr
 License: \tab GPL-2\cr
 }
}
\section{asymLD functions}{

The function compute.ALD() calculates asymmetric LD for haplotype frequency data.
The function compute.AShomz() calculates allele specific homozygosity values for haplotype frequency data.
}
\references{
Thomson G, Single RM. Conditional Asymmetric Linkage Disequilibrium (ALD): Extending the Bi-Allelic r^2 Measure. Genetics. 2014  198(1):321-331. PMID:25023400
}
\keyword{LD}
\keyword{disequilibrium}
\keyword{genetics}
\keyword{linkage}

