% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{hla.freqs}
\alias{hla.freqs}
\title{Pairwise haplotype frequencies for 7 HLA loci}
\format{A data frame with 3063 observations on the following 5 variables.
 \describe{
\item{\code{haplo.freq}}{a numeric vector}
\item{\code{locus1}}{a character vector}
\item{\code{locus2}}{a character vector}
\item{\code{allele1}}{a character vector}
\item{\code{allele2}}{a character vector}
}}
\source{
Wilson, C., 2010 Identifying polymorphisms associated with risk for the development of myopericarditis following smallpox vaccine. The Immunology Database and Analysis Portal (ImmPort), Study #26.
}
\usage{
data(hla.freqs)
}
\description{
HLA haplotype frequencies for 21 pairs of HLA loci in a set of 300 controls
from a study of myopericarditis incidence following smallpox vaccination.
}
\references{
https://immport.niaid.nih.gov/immportWeb/clinical/study/displayStudyDetails.do?itemList=SDY26
}
\keyword{data}

