% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{snp.freqs}
\alias{snp.freqs}
\title{Pairwise haplotype frequencies for 4 SNP loci}
\format{A data frame with 20 observations on the following 5 variables.
  \describe{
    \item{\code{locus1}}{a character vector}
    \item{\code{locus2}}{a character vector}
    \item{\code{allele1}}{a character vector}
    \item{\code{allele2}}{a character vector}
    \item{\code{haplo.freq}}{a numeric vector}
  }}
\source{
de Bakker, P. I., G. McVean, P. C. Sabeti, M. M. Miretti, T. Green et al., 2006 A high-resolution HLA and SNP haplotype map for disease association studies in the extended human MHC. Nat.Genet. 38: 1166-1172.
}
\usage{
data(snp.freqs)
}
\description{
HLA haplotype frequencies for 6 pairs of SNP loci from the de Bakker et al. 2006 data for 90 unrelated individuals with European ancestry (CEU) from the Centre d'Etude du Polymorphisme Humain (CEPH)
collection obtained from the Tagger/MHC webpage.
}
\details{
For bi-allelic SNP data the ALD measures are symmetric and equivalent to the r correlation measure of LD.
}
\references{
http://www.broadinstitute.org/mpg/tagger/mhc.html
}
\keyword{data}

